Move common views methods into their own file
This file is meant to be used by the views only, nothing else.
